backend/requirements.txt
# Core ADK framework
google-adk==1.13.0

# For configuration management
pydantic-settings==2.10.1

# For structured logging
structlog==25.4.0

# For interacting with Google Cloud Storage, BigQuery, Gemini, and Firestore
google-cloud-storage==2.19.0
google-genai==1.32.0
google-cloud-bigquery==3.37.0
google-cloud-tasks==2.19.3
google-cloud-firestore==2.21.0
firebase-admin==6.9.0

# For making HTTP requests to external APIs (ClinVar, gnomAD)
httpx==0.28.1

# Web server used by ADK
uvicorn==0.35.0